Paste URL/regex selection to prompt if key is uppercase.
Fixes #1975.

In copy-regex/show-urls-copy mode, if the last input character
was uppercase, copy the selection to the clipboard _and_ paste
it. This is useful for taking a file path from a command output
(log, git, test failure, etc.) and using it in another command.

This is inspired by the behavior of copy mode in wezterm:
https://wezterm.org/quickselect.html

I could have made it check every character in the hint, but it
seemed fine to assume that if the last character was uppercase,
the user wanted this behavior.